Refactor CI workflows and apply RxCpp patch for modern compilers.
Update GitHub Actions to use updated checkout actions and a more efficient dependency management strategy. Add a patch to fix RxCpp compatibility with modern compilers and perform minor code cleanups in the CNC module.

+++ ralgo/cnc/interpreter.cpp
@@ -572,7 +572,7 @@ namespace cnc
         cleanup();
 
         // 5. Reset positions to zero
-        for (size_t i = 0; i < total_axes; i++)
+        for (int i = 0; i < total_axes; i++)
         {
             _final_position[i] = 0;
             revolver->get_steppers()[i]->set_counter_value(0);
